split querybuilder
